File and directory restructure
